requirements.txt
# Exam Buddy AI Pro - Enhanced Dependencies (With Advanced ML Features)
# Install with: pip install -r requirements.txt

# Core Framework
streamlit>=1.28.0

# AI/ML
groq
numpy>=1.24.0
scipy>=1.10.0

# Data Visualization
plotly>=5.17.0

# Data Processing
pandas>=2.0.0

# ==========================================
# NEW DEPENDENCIES FOR ENHANCED FEATURES
# ==========================================

# Audio Processing (Voice Search)
SpeechRecognition>=3.10.0

# PDF Processing (File Upload)
PyPDF2>=3.0.0

# Image Processing (OCR for images)
pytesseract>=0.3.10
Pillow>=10.0.0

# ==========================================
# Optional but recommended
# ==========================================
python-dateutil>=2.8.2
scikit-learn>=1.3.0

# ==========================================
# SPECIAL INSTALLATION NOTES
# ==========================================

# PyAudio (for voice input) - Install separately:
# 
# Windows:
#   pip install pipwin
#   pipwin install pyaudio
#
# Mac:
#   brew install portaudio
#   pip install pyaudio
#
# Linux:
#   sudo apt-get install python3-pyaudio
#   pip install pyaudio

# Tesseract OCR (for image text extraction) - Install separately:
#
# This is NOT a Python package - install from:
#
# Windows: 
#   Download from: https://github.com/UB-Mannheim/tesseract/wiki
#   Install to: C:\Program Files\Tesseract-OCR
#   Add to PATH
#
# Mac:
#   brew install tesseract
#
# Linux (Ubuntu/Debian):
#   sudo apt-get update
#   sudo apt-get install tesseract-ocr
#   sudo apt-get install libtesseract-dev